+++ db/DBScript.sql
@@ -4,8 +4,9 @@ DROP DATABASE IF EXISTS `BookInventory`;
 CREATE DATABASE `BookInventory`;
 
 USE `BookInventory`;
+
 CREATE TABLE `Book` (\
-  `id` int(11) NOT NULL,\
+  `id` bigint(13) NOT NULL,\
   `name` varchar(45) NOT NULL,\
   `author` varchar(45) NOT NULL,\
   `page` int(11) NOT NULL,\
